+ 86cd07f60220b9371b7815e662470a79a29fd21e Add some checks around wmf_civicrm_update_8025() to make it work locally on vagrant.
+ 5651cf78d6c00d25450d97197a77d35410cea20f Recurring QC: always set installments to 0 on new subs
+ e2602b271518e6bd6402051d51fdd5d2d7f734c1 Convert Civi-errors in munge to WMFExceptions

+++ sites/all/modules/offline2civicrm/ChecksFile.php
@@ -750,11 +750,12 @@ abstract class ChecksFile {
    * Get the ID of our anonymous contact.
    *
    * @return int|NULL
+   * @throws \CiviCRM_API3_Exception
    */
   protected function getAnonymousContactID() {
     static $contactID = NULL;
     if (!$contactID) {
-      $contactID = civicrm_api3('Contact', 'getvalue', array(
+      $contactID = (int) civicrm_api3('Contact', 'getvalue', array(
         'return' => 'id',
         'contact_type' => 'Individual',
         'first_name' => 'Anonymous',
